Multi_Container_LangChain/rag_chatbot/requirements.txt
fastapi == 0.115.0
uvicorn == 0.30.6
langchain == 0.3.1
langchain_groq
httpx == 0.27.2
python-dotenv ==1.0.1
pytest == 8.3.2
langchain_huggingface
langchain_community
langchain_core
sentence-transformers==2.2.2

